+++ Models/Order.cs
@@ -21,7 +21,7 @@ namespace Antique.Models
         public string DeliveryNum { get; set; }
         public double TotalPrice { get; set; }
         public bool isDone { get; set; }
-        public double Invoice { get; set; }
+        public long Invoice { get; set; }
         public ICollection<OrderItem> Items { get; set; }
     }
 }
