+++ ViewModel/InvoiceModel.cs
@@ -8,6 +8,6 @@ namespace Antique.ViewModel
     public class InvoiceModel
     {
         public int ID { get; set; }
-        public int Invoice { get; set; }
+        public long Invoice { get; set; }
     }
 }
